refactor: transform operator stats resolution into class with attribute cache
1. Renamed update_operator_usage_stats.py -> stats_resolver.py.
2. Created new class OperatorUsageStatsResolver	with attribute _cache of type CachedContext.
3. Transformed all of the old module functions into methods of the new class.
4. Initialized the class in main.
5. Updated tests to reflect these changes.
