+++ apps/worker/tests/test_stats_resolver.py
@@ -2,14 +2,19 @@ import pytest
 from pytest_mock import MockerFixture
 from pytest import CaptureFixture
 from datetime import date
-from typing import List, Dict
+from typing import List
 
-from pullsar import update_operator_usage_stats as stats
+from pullsar.stats_resolver import OperatorUsageStatsResolver
 from pullsar.operator_bundle_model import OperatorBundle
 from pullsar.quay_client import QuayClient
 from pullsar.pyxis_client import PyxisClient
 from pullsar.parse_operators_catalog import RepositoryMap
-from pullsar.cached_context import CachedContext
+
+
+@pytest.fixture
+def stats() -> OperatorUsageStatsResolver:
+    """Provides an OperatorUsageStatsResolver object."""
+    return OperatorUsageStatsResolver()
 
 
 @pytest.fixture
@@ -21,7 +26,7 @@ def sample_bundles() -> list[OperatorBundle]:
     return [b1, b2, b3]
 
 
-def test_tag_in_tag_map() -> None:
+def test_tag_in_tag_map(stats: OperatorUsageStatsResolver) -> None:
     """Tests the tag equivalence logic."""
     tag_map = {
         "v1.0": OperatorBundle("op.v1.0", "op", "img"),
@@ -35,7 +40,9 @@ def test_tag_in_tag_map() -> None:
     assert stats.tag_in_tag_map("v2.0.0", tag_map) == "2.0.0"
 
 
-def test_create_local_tag_digest_maps(sample_bundles: list[OperatorBundle]) -> None:
+def test_create_local_tag_digest_maps(
+    stats: OperatorUsageStatsResolver, sample_bundles: list[OperatorBundle]
+) -> None:
     """Tests the creation of local tag and digest maps."""
     tag_map, digest_map = stats.create_local_tag_digest_maps(sample_bundles)
     assert len(tag_map) == 3
@@ -47,13 +54,13 @@ def test_create_local_tag_digest_maps(sample_bundles: list[OperatorBundle]) -> N
     assert digest_map.get("sha256:abc") == sample_bundles[2]
 
 
-def test_extract_date() -> None:
+def test_extract_date(stats: OperatorUsageStatsResolver) -> None:
     """Tests the date extraction from Quay's log format."""
     datetime_str = "Mon, 14 Jul 2025 16:23:18 -0000"
     assert stats.extract_date(datetime_str) == date(2025, 7, 14)
 
 
-def test_filter_pull_repo_logs() -> None:
+def test_filter_pull_repo_logs(stats: OperatorUsageStatsResolver) -> None:
     """Tests that 'pull_repo' logs are filtered and simplified."""
     quay_logs = [
         {"kind": "push_repo", "datetime": "Mon, 14 Jul 2025 10:00:00 -0000"},
@@ -92,7 +99,9 @@ def bundles_to_translate() -> List[OperatorBundle]:
 
 
 def test_resolve_repositories_success_with_multiple_bundles(
-    mocker: MockerFixture, bundles_to_translate: List[OperatorBundle]
+    mocker: MockerFixture,
+    stats: OperatorUsageStatsResolver,
+    bundles_to_translate: List[OperatorBundle],
 ) -> None:
     """
     Tests that multiple non-Quay bundles from different repositories are
@@ -127,10 +136,11 @@ def test_resolve_repositories_success_with_multiple_bundles(
         "connect-org-b/repo-b": [bundles_to_translate[1]],
     }
     quay_map: RepositoryMap = {}
-    known_images_map: Dict[str, str] = {}
 
     stats.resolve_not_quay_repositories(
-        mock_pyxis_client, not_quay_map, quay_map, known_images_map, {}
+        mock_pyxis_client,
+        not_quay_map,
+        quay_map,
     )
 
     assert len(quay_map) == 2
@@ -145,23 +155,29 @@ def test_resolve_repositories_success_with_multiple_bundles(
     assert bundle_b.name == "op-b.v1"
     assert bundle_b.image == "quay.io/quay-org-b/repo-b@sha256:digest2"
 
+    known_images_map = stats._cache.known_image_translations
     assert len(known_images_map) == 2
     assert known_images_map[bundles_to_translate[0].image] == bundle_a.image
     assert known_images_map[bundles_to_translate[1].image] == bundle_b.image
 
 
 def test_update_image_digests(
-    mocker: MockerFixture, sample_bundles: list[OperatorBundle]
+    mocker: MockerFixture,
+    stats: OperatorUsageStatsResolver,
+    sample_bundles: list[OperatorBundle],
 ) -> None:
     """Tests that image digests are correctly updated from mocked Quay tags."""
-    mock_quay_client = mocker.Mock(spec=QuayClient)
-    mock_quay_client.get_repo_tags.return_value = [
+    tags = [
         {"name": "1", "manifest_digest": "sha256:digest_for_v1"},
         {"name": "v2", "manifest_digest": "sha256:digest_for_v2"},
     ]
-    repo_map = {"org/repo": sample_bundles}
+    repo = "org/repo"
 
-    stats.update_image_digests(mock_quay_client, repo_map, {})
+    mock_quay_client = mocker.Mock(spec=QuayClient)
+    mock_quay_client.get_repo_tags.return_value = tags
+    repo_map = {repo: sample_bundles}
+
+    stats.update_image_digests(mock_quay_client, repo_map)
 
     # digests were updated
     assert sample_bundles[0].digest == "sha256:digest_for_v1"
@@ -170,9 +186,19 @@ def test_update_image_digests(
     assert sample_bundles[2].digest == "sha256:abc"
     mock_quay_client.get_repo_tags.assert_called_once_with("org/repo")
 
+    cached_repo_path_to_tags = stats._cache.repo_path_to_tags
+    assert len(cached_repo_path_to_tags) == 1
+
+    cached_repo_tags = cached_repo_path_to_tags[repo]
+    assert len(cached_repo_tags) == 2
+    assert tags[0] in cached_repo_tags
+    assert tags[1] in cached_repo_tags
+
 
 def test_update_image_pull_counts(
-    mocker: MockerFixture, sample_bundles: list[OperatorBundle]
+    mocker: MockerFixture,
+    stats: OperatorUsageStatsResolver,
+    sample_bundles: list[OperatorBundle],
 ) -> None:
     """Tests that pull counts are correctly retrieved from Quay logs."""
     mock_quay_client = mocker.Mock(spec=QuayClient)
@@ -198,21 +224,28 @@ def test_update_image_pull_counts(
             "metadata": {"manifest_digest": "sha256:abc"},
         },
     ]
-    mock_quay_client.get_repo_logs.return_value = quay_logs
-    repo_map = {"org/repo": sample_bundles}
+    repo = "org/repo"
 
-    stats.update_image_pull_counts(
-        mock_quay_client, repo_map, log_days=7, repo_path_to_logs={}
-    )
+    mock_quay_client.get_repo_logs.return_value = quay_logs
+    repo_map = {repo: sample_bundles}
+
+    stats.update_image_pull_counts(mock_quay_client, repo_map, log_days=7)
 
     assert sample_bundles[0].pull_count == {date(2025, 7, 14): 2}
     assert sample_bundles[1].pull_count == {}
     assert sample_bundles[2].pull_count == {date(2025, 7, 15): 1}
-    mock_quay_client.get_repo_logs.assert_called_once_with("org/repo", 7)
+    mock_quay_client.get_repo_logs.assert_called_once_with(repo, 7)
+
+    cached_repo_path_to_logs = stats._cache.repo_path_to_logs
+    assert len(cached_repo_path_to_logs) == 1
+
+    cached_repo_logs = cached_repo_path_to_logs[repo]
+    assert len(cached_repo_logs) == 3
 
 
 def test_update_image_pull_counts_no_logs(
     mocker: MockerFixture,
+    stats: OperatorUsageStatsResolver,
     sample_bundles: list[OperatorBundle],
 ) -> None:
     """
@@ -221,18 +254,23 @@ def test_update_image_pull_counts_no_logs(
     """
     mock_quay_client = mocker.Mock(spec=QuayClient)
     mock_quay_client.get_repo_logs.return_value = []
-    repo_map = {"org/repo": sample_bundles}
+    repo = "org/repo"
+    repo_map = {repo: sample_bundles}
 
-    stats.update_image_pull_counts(
-        mock_quay_client, repo_map, log_days=7, repo_path_to_logs={}
-    )
+    stats.update_image_pull_counts(mock_quay_client, repo_map, log_days=7)
 
     for bundle in sample_bundles:
         assert bundle.pull_count == {}
 
+    cached_repo_path_to_logs = stats._cache.repo_path_to_logs
+    assert len(cached_repo_path_to_logs) == 1
+    assert len(cached_repo_path_to_logs[repo]) == 0
+
 
 def test_print_operator_usage_stats(
-    capsys: CaptureFixture[str], sample_bundles: list[OperatorBundle]
+    capsys: CaptureFixture[str],
+    stats: OperatorUsageStatsResolver,
+    sample_bundles: list[OperatorBundle],
 ) -> None:
     """Tests that only bundles with pull counts are printed."""
     sample_bundles[0].pull_count[date(2025, 7, 14)] = 5
@@ -245,25 +283,29 @@ def test_print_operator_usage_stats(
     assert "op.v2" not in captured.out
 
 
-def test_update_operator_usage_stats_flow(mocker: MockerFixture) -> None:
+def test_update_operator_usage_stats_flow(
+    mocker: MockerFixture, stats: OperatorUsageStatsResolver
+) -> None:
     """
     Tests the main orchestration function, mocking its dependencies.
     """
-    mock_render = mocker.patch(
-        "pullsar.update_operator_usage_stats.render_operator_catalog"
-    )
+    mock_render = mocker.patch("pullsar.stats_resolver.render_operator_catalog")
     mock_create_maps = mocker.patch(
-        "pullsar.update_operator_usage_stats.create_repository_paths_maps",
+        "pullsar.stats_resolver.create_repository_paths_maps",
         return_value=({"repo": []}, {"repo": []}, {}),
     )
+
+    mock_resolve_repos = mocker.patch(
+        "pullsar.stats_resolver.OperatorUsageStatsResolver.resolve_not_quay_repositories"
+    )
     mock_update_digests = mocker.patch(
-        "pullsar.update_operator_usage_stats.update_image_digests"
+        "pullsar.stats_resolver.OperatorUsageStatsResolver.update_image_digests"
     )
     mock_update_pulls = mocker.patch(
-        "pullsar.update_operator_usage_stats.update_image_pull_counts"
+        "pullsar.stats_resolver.OperatorUsageStatsResolver.update_image_pull_counts"
     )
     mock_print_stats = mocker.patch(
-        "pullsar.update_operator_usage_stats.print_operator_usage_stats"
+        "pullsar.stats_resolver.OperatorUsageStatsResolver.print_operator_usage_stats"
     )
     mock_quay_client = mocker.Mock(spec=QuayClient)
     mock_pyxis_client = mocker.Mock(spec=PyxisClient)
@@ -272,13 +314,13 @@ def test_update_operator_usage_stats_flow(mocker: MockerFixture) -> None:
     stats.update_operator_usage_stats(
         quay_client=mock_quay_client,
         pyxis_client=mock_pyxis_client,
-        cache=CachedContext(),
         log_days=7,
         catalog_image="my-image:latest",
     )
 
     mock_render.assert_called_once()
     mock_create_maps.assert_called_once()
+    mock_resolve_repos.assert_called_once()
     mock_update_digests.assert_called_once()
     mock_update_pulls.assert_called_once()
     mock_print_stats.assert_called_once()
